+++ Interface/AddOns/PanzaUI/core.lua
@@ -418,7 +418,7 @@ end
 
 local function configMapElements()
   -- Set Alpha and Scale
-  SetScaleForFrame(MinimapZoneText, 1.2)
+  SetScaleForFrame(MinimapZoneText, 1.3)
   SetScaleForFrame(Minimap, 1.0)
   SetScaleForFrame(MinimapCluster, 1.0)
   SetAlphaForFrame(MinimapCluster.BorderTop, 0)
@@ -427,6 +427,11 @@ local function configMapElements()
   HideFrameAndUnregisterEvents(Minimap.ZoomIn)
   HideFrameAndUnregisterEvents(Minimap.ZoomOut)
   HideFrameAndUnregisterEvents(MinimapCluster.TrackingFrame)
+
+  -- Set Scale to Clock button
+  LoadAddOn("Blizzard_TimeManager")
+  local region = TimeManagerClockButton:GetRegions()
+  TimeManagerClockButton:SetScale(1.3)
 
   -- Set Garrison button scale
   SetScaleForFrame(ExpansionLandingPageMinimapButton, 0.85)